requirements.txt
asgiref==3.2.7
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.0.7
djangorestframework==3.11.0
ecdsa==0.15
html5lib==1.0.1
idna==2.9
Pillow==7.1.2
pyasn1==0.4.8
pycparser==2.20
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
urllib3==1.25.9
WeasyPrint==51
webencodings==0.5.1
